Add "prepare" to package.json
See: https://stackoverflow.com/a/57829251/15187023

Without this, new dist folder is not installed on npm install. 
So for instance, "Add duration property on web #22" was not part of the build after npm install, since it would only be build npm publish.

I am no expert in this matter, so I am not sure if "prepublishOnly" should be removed.